moved legal hint to bottom
