+++ app/components/Common/Header/MainScreenHeader.js
@@ -1,22 +1,12 @@
 import React from 'react';
 import { View, Text, StyleSheet, TouchableOpacity, Image } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
 import { Colors, Typography } from '_styles';
 import i18next from '../../../languages/languages';
 
 const MainScreenHeader = ({ onPressLogin, isUserLogin, accessibilityLabel, testID, photo }) => {
-  const navigation = useNavigation();
-  const onPressLegals = () => {
-    navigation.navigate('Legals');
-  };
-
   return (
     <View style={styles.container}>
-      {!isUserLogin ? (<View>
-        <Text onPress={onPressLegals} style={styles.legalText}>
-          {i18next.t('label.legal_docs')}
-        </Text>
-        </View>) : <View/> }
+      <View />
       <TouchableOpacity
         onPress={onPressLogin}
         testID={testID}
@@ -45,11 +35,6 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     justifyContent: 'space-between',
     flex: 1,
-  },
-  legalText: {
-    color: Colors.TEXT_COLOR,
-    fontSize: Typography.FONT_SIZE_10,
-    fontFamily: Typography.FONT_FAMILY_REGULAR,
   },
   loginText: {
     color: Colors.PRIMARY,
